requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
pymongo
bcrypt
python-dotenv